# docs/swagger.yaml
basePath: /api
definitions:
  handler.contactInfoResponse:
    properties:
      info:
        $ref: '#/definitions/whatsapp.ContactInfo'
    type: object
  handler.getCheckPhonesBody:
    properties:
      phones:
        items:
          type: string
        type: array
    type: object
  handler.getCheckPhonesResponse:
    properties:
      phones:
        items:
          $ref: '#/definitions/whatsapp.IsOnWhatsAppResponse'
        type: array
    type: object
  handler.getMessagesBody:
    properties:
      phone:
        type: string
    type: object
  handler.getMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/response.Message'
        type: array
    type: object
  handler.getProfileInfoResponse:
    properties:
      info:
        $ref: '#/definitions/whatsapp.ContactInfo'
    type: object
  handler.getQrCodeResponse:
    properties:
      qrcode:
        type: string
    type: object
  handler.getStatusResponse:
    properties:
      status:
        type: string
    type: object
...
  handler.proxyRequest:
    properties:
      proxy:
        type: string
    type: object
  handler.sendAudioMessageBody:
    properties:
      base64:
        type: string
      phone:
        type: string
    type: object
  handler.sendAudioMessageResponse:
    properties:
      message:
        $ref: '#/definitions/response.Message'
    type: object
...
  response.Message:
    properties:
      body:
        type: string
      chat:
        type: string
      from_me:
        type: boolean
      id:
        type: integer
      media_base64:
        type: string
      media_mimetype:
        type: string
      media_type:
        type: string
      message_id:
        type: string
      sender:
        type: string
      timestamp:
        type: string
    type: object
  whatsapp.ContactInfo:
    properties:
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      status:
        type: string
    type: object
  whatsapp.IsOnWhatsAppResponse:
    properties:
      is_registered:
        type: boolean
      phone:
        type: string
      query:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API to handle multiple WhatsApp instances
  title: ZapMeow API
  version: "1.0"
paths:
  /{instanceId}/chat/messages:
    post:
      consumes:
      - application/json
      description: Returns chat messages from the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Phone
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.getMessagesBody'
      produces:
      - application/json
      responses:
        "200":
          description: List of chat messages
          schema:
            $ref: '#/definitions/handler.getMessagesResponse'
      summary: Get WhatsApp Chat Messages
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/audio:
    post:
      consumes:
      - application/json
      description: Sends an audio message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Audio message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.sendAudioMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/handler.sendAudioMessageResponse'
      summary: Send Audio Message on WhatsApp
      tags:
      - WhatsApp Chat
...
  /{instanceId}/check/phones:
    post:
      consumes:
      - application/json
      description: Verifies if the phone numbers in the provided list are registered
        WhatsApp users.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Phone list
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.getCheckPhonesBody'
      produces:
      - application/json
      responses:
        "200":
          description: List of verified numbers
          schema:
            $ref: '#/definitions/handler.getCheckPhonesResponse'
      summary: Check Phones on WhatsApp
      tags:
      - WhatsApp Phone Verification
  /{instanceId}/contact/info:
    get:
      consumes:
      - application/json
      description: Retrieves contact information.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Phone
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Contact Information
          schema:
            $ref: '#/definitions/handler.contactInfoResponse'
      summary: Get Contact Information
      tags:
      - WhatsApp Contact
  /{instanceId}/logout:
    post:
      consumes:
      - application/json
      description: Logs out from the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
      summary: Logout from WhatsApp
      tags:
      - WhatsApp Logout
  /{instanceId}/profile:
    get:
      consumes:
      - application/json
      description: Retrieves profile information.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile Information
          schema:
            $ref: '#/definitions/handler.getProfileInfoResponse'
      summary: Get Profile Information
      tags:
      - WhatsApp Profile
  /{instanceId}/qrcode:
    post:
      consumes:
      - application/json
      description: Returns a QR code to initiate WhatsApp login.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Text message body
        in: body
        name: data
        schema:
          $ref: '#/definitions/handler.proxyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: QR Code
          schema:
            $ref: '#/definitions/handler.getQrCodeResponse'
      summary: Get WhatsApp QR Code
      tags:
      - WhatsApp Login
  /{instanceId}/status:
    get:
      consumes:
      - application/json
      description: Returns the status of the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status Response
          schema:
            $ref: '#/definitions/handler.getStatusResponse'
      summary: Get WhatsApp Instance Status
      tags:
      - WhatsApp Status
...
swagger: "2.0"